requirements.txt
alembic==1.15.1
bcrypt==4.3.0
blinker==1.9.0
boto3==1.37.9
botocore==1.37.9
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3
Werkzeug
